requirement.txt
cryptography==2.3.1
Django==2.1.1
djangorestframework==3.8.2
dnspython==1.15.0
fabric==2.4.0
lxml==4.2.5
mysqlclient==1.3.13
pybindxml==0.5
